Reuse display line when adding tokens
