+++ packages/browser-wallet/src/popup/pages/AddTokens/ExternalAddTokens.tsx
@@ -1,4 +1,3 @@
-import clsx from 'clsx';
 import React, { useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useAtom, useAtomValue, useSetAtom } from 'jotai';
@@ -9,75 +8,15 @@ import { addToastAtom } from '@popup/state';
 import ExternalRequestLayout from '@popup/page-layouts/ExternalRequestLayout';
 import { currentAccountTokensAtom } from '@popup/store/token';
 import { useAsyncMemo } from 'wallet-common-helpers';
-import { NetworkConfiguration, TokenIdAndMetadata, TokenMetadata } from '@shared/storage/types';
+import { NetworkConfiguration, TokenIdAndMetadata } from '@shared/storage/types';
 import { ContractDetails, fetchContractName, getTokenMetadata, getTokenUrl } from '@shared/utils/token-helpers';
 import { jsonRpcClientAtom, networkConfigurationAtom } from '@popup/store/settings';
 import { selectedAccountAtom } from '@popup/store/account';
 import { fullscreenPromptContext } from '@popup/page-layouts/FullscreenPromptLayout';
 import { Input as UncontrolledInput } from '@popup/shared/Form/Input';
 import Button from '@popup/shared/Button';
-import { Checkbox } from '@popup/shared/Form/Checkbox';
 import { isHex, JsonRpcClient } from '@concordium/web-sdk';
-import TokenBalance from '@popup/shared/TokenBalance';
-
-enum ChoiceStatus {
-    discarded,
-    chosen,
-    existing,
-}
-
-/**
- * Displays a CIS-2 Token
- */
-export function DisplayToken({
-    status,
-    metadata,
-    onClick,
-    balance,
-    updateStatus,
-}: {
-    status: ChoiceStatus;
-    metadata: TokenMetadata;
-    onClick?: () => void;
-    balance: bigint;
-    updateStatus: () => void;
-}) {
-    const clickable = status !== ChoiceStatus.existing;
-    return (
-        <Button clear className="add-tokens-list__token" title={metadata.description} onClick={onClick}>
-            <div className="flex align-center h-full">
-                <img
-                    className="add-tokens-prompt__token-display"
-                    src={metadata.thumbnail?.url}
-                    alt={metadata.name ?? ''}
-                />
-                <div>
-                    {metadata.name}
-                    <div
-                        className={clsx(
-                            'add-tokens-list__token-balance',
-                            balance !== 0n && 'add-tokens-list__token-balance--owns'
-                        )}
-                    >
-                        Your balance:
-                        <TokenBalance balance={balance} decimals={metadata.decimals ?? 0} symbol={metadata.symbol} />
-                    </div>
-                </div>
-            </div>
-            {clickable && (
-                <Checkbox
-                    onClick={(e) => {
-                        e.stopPropagation();
-                    }}
-                    onChange={updateStatus}
-                    className="add-tokens-prompt__checkbox"
-                    checked={status === ChoiceStatus.chosen}
-                />
-            )}
-            {clickable || <p className="m-l-auto text-faded">Already Added</p>}
-        </Button>
-    );
-}
+import ContractTokenLine, { ChoiceStatus } from '@popup/shared/ContractTokenLine';
 
 type Props = {
     respond(ids: string[]): void;
@@ -192,12 +131,12 @@ export default function SignMessage({ respond }: Props) {
                 />
                 <div className="add-tokens-prompt__token-container w-full">
                     {addingTokens.map((token, index) => (
-                        <DisplayToken
+                        <ContractTokenLine
                             key={token.id}
                             status={token.status}
-                            metadata={token.metadata}
-                            balance={0n}
-                            updateStatus={() => updateTokenStatus(index)}
+                            token={{ ...token, balance: 0n }}
+                            onToggleChecked={() => updateTokenStatus(index)}
+                            onClick={() => {}}
                         />
                     ))}
                 </div>
