+++ packages/browser-wallet/src/popup/shared/ContractTokenLine/ContractTokenLine.scss
@@ -1,0 +1,44 @@
+$element-border-radius: rem(5px);
+
+.contract-token-line {
+    &__token {
+        padding: rem(5px) rem(10px);
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        cursor: pointer;
+        height: rem(42px);
+        text-align: left;
+        border-bottom: rem(1px) solid $color-lines;
+
+        &:last-of-type {
+            border-bottom-color: transparent;
+        }
+
+        &:first-child {
+            border-top-left-radius: $element-border-radius;
+            border-top-right-radius: $element-border-radius;
+        }
+
+        img {
+            max-height: 100%;
+            width: rem(40px);
+            object-position: center;
+            object-fit: contain;
+            margin-right: rem(5px);
+        }
+    }
+
+    &__token-balance {
+        font-size: 10px;
+        color: $color-text-faded;
+
+        &--owns {
+            color: $color-cta;
+        }
+    }
+
+    &__checkbox {
+        width: rem(14px);
+    }
+}
