+++ packages/browser-wallet/src/popup/styles/_components.scss
@@ -17,6 +17,7 @@
 @import '../shared/Toast/Toast';
 @import '../shared/TabBar/TabBar';
 @import '../shared/TokenDetails/TokenDetails';
+@import '../shared/ContractTokenLine/ContractTokenLine';
 
 // Pages
 @import '../pages/Account';
